rename TripsContainer to ViewTripContainer
