+++ src/styled_scenes/EditTrip/components/ToolBar/OwnerToolBar.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import EditTripForm from './EditTripForm';
-import EditTripContainer from 'scenes/trips/containers/edit_trip_container';
+import EditTripContainer from 'scenes/trips/containers/EditTripContainer';
 import { media } from 'libs/styled';
 
 const Wrap = styled.div`
